+++ src/components/BottomSheet.tsx
@@ -7,6 +7,7 @@ import {
   type MouseEvent,
 } from "react"
 import {Close} from "@/assets/icons"
+import useBodyScrollLock from "@/hooks/use-body-scroll-lock"
 
 type TProps = {
   children: ReactNode
@@ -26,6 +27,8 @@ const BottomSheet = ({
   isControllingScroll = true,
   ...props
 }: TProps) => {
+  const {lockScroll, openScroll} = useBodyScrollLock()
+
   useEffect(() => {
     if (!isHandlingHistory) {
       return
@@ -43,31 +46,12 @@ const BottomSheet = ({
       return
     }
 
-    let scrollPosition = 0
-    const lockScroll = () => {
-      // for IOS safari
-      scrollPosition = window.pageYOffset
-      document.body.style.overflow = "hidden"
-      document.body.style.position = "fixed"
-      document.body.style.top = `-${scrollPosition}px`
-      document.body.style.width = "100%"
-    }
-
-    const openScroll = () => {
-      // for IOS safari
-      document.body.style.removeProperty("overflow")
-      document.body.style.removeProperty("position")
-      document.body.style.removeProperty("top")
-      document.body.style.removeProperty("width")
-      window.scrollTo(0, scrollPosition)
-    }
-
     if (isOpen) {
       lockScroll()
     } else {
       openScroll()
     }
-  }, [isControllingScroll, isOpen])
+  }, [isControllingScroll, isOpen, lockScroll, openScroll])
 
   function handlePropagation(event: MouseEvent<HTMLDivElement>) {
     event.stopPropagation()
